+++ src/main/kotlin/com/atlassian/performance/tools/virtualusers/api/config/VirtualUserBehavior.kt
@@ -10,6 +10,8 @@ import com.atlassian.performance.tools.virtualusers.api.users.UserGenerator
 import com.atlassian.performance.tools.virtualusers.logs.LogConfiguration
 import org.apache.logging.log4j.core.config.AbstractConfiguration
 import java.time.Duration
+import java.nio.file.Path
+import java.nio.file.Paths
 
 /**
  * @param [maxOverhead] Maximum time to be running, but not applying load.
@@ -19,6 +21,7 @@ class VirtualUserBehavior private constructor(
         message = "There should be no need to display help from Java API. Read the Javadoc or sources instead."
     )
     internal val help: Boolean,
+    internal val results: Path,
     internal val scenario: Class<out Scenario>,
     val load: VirtualUserLoad,
     val maxOverhead: Duration,
@@ -43,6 +46,7 @@ class VirtualUserBehavior private constructor(
         skipSetup: Boolean
     ) : this(
         help = help,
+        results = Paths.get("."),
         scenario = scenario,
         load = load,
         maxOverhead = Duration.ofMinutes(5),
@@ -108,6 +112,7 @@ class VirtualUserBehavior private constructor(
     class Builder(
         private var scenario: Class<out Scenario>
     ) {
+        private var results: Path = Paths.get(".")
         private var load: VirtualUserLoad = VirtualUserLoad.Builder().build()
         private var maxOverhead: Duration = Duration.ofMinutes(5)
         private var seed: Long = 12345
@@ -117,6 +122,7 @@ class VirtualUserBehavior private constructor(
         private var skipSetup = false
         private var userGenerator: Class<out UserGenerator> = SuppliedUserGenerator::class.java
 
+        fun results(results: Path) = apply { this.results = results }
         fun scenario(scenario: Class<out Scenario>) = apply { this.scenario = scenario }
         fun load(load: VirtualUserLoad) = apply { this.load = load }
         fun maxOverhead(maxOverhead: Duration) = apply { this.maxOverhead = maxOverhead }
@@ -155,6 +161,7 @@ class VirtualUserBehavior private constructor(
         @Suppress("DEPRECATION")
         fun build(): VirtualUserBehavior = VirtualUserBehavior(
             help = false,
+            results = results,
             scenario = scenario,
             load = load,
             maxOverhead = maxOverhead,