+++ src/test/kotlin/com/atlassian/performance/tools/virtualusers/api/EntryPointIT.kt
@@ -8,9 +8,12 @@ import com.atlassian.performance.tools.virtualusers.ChromeContainer
 import com.atlassian.performance.tools.virtualusers.DockerJira
 import com.atlassian.performance.tools.virtualusers.SimpleScenario
 import com.atlassian.performance.tools.virtualusers.lib.infrastructure.Jperf425WorkaroundMysqlDatabase
+import org.assertj.core.api.Assertions.assertThat
 import org.junit.Test
 import java.net.URI
+import java.nio.file.Paths
 import java.time.Duration
+import java.util.*
 
 class EntryPointIT {
 
@@ -35,6 +38,10 @@ class EntryPointIT {
 
     @Test
     fun shouldRunWith3_2_0_Args() {
+        val resultPath = Paths.get("build")
+            .resolve("vu-node-result")
+            .resolve(UUID.randomUUID().toString())
+        val desiredTotalTime = Duration.ofMinutes(2)
         DockerJira(smallDataset).runWithJira { jira: URI ->
             com.atlassian.performance.tools.virtualusers.api.main(arrayOf(
                 "--jira-address", jira.toString(),
@@ -43,12 +50,21 @@ class EntryPointIT {
                 "--virtual-users", "1",
                 "--hold", "PT0S",
                 "--ramp", "PT0S",
-                "--flat", "PT2M",
+                "--flat", desiredTotalTime.toString(),
                 "--scenario", SimpleScenario::class.java.name,
                 "--browser", ChromeContainer::class.java.name,
-                "--diagnostics-limit", "64",
-                "--seed", "-9183767962456348780"
+                "--diagnostics-limit", "3",
+                "--seed", "-9183767962456348780",
+                "--result", resultPath.toString()
             ))
         }
+
+        val vuResult = VirtualUserNodeResult(resultPath)
+            .listResults()
+            .last()
+        val scenarioLabels = vuResult
+            .streamScenarioMetrics()
+            .map { it.label }
+        assertThat(scenarioLabels).containsOnly("Log In", "View Issue")
     }
 }